added edit and delete to my-reviews

+++ elteck-angular-final/src/app/elteck/review-page/review-page.component.ts
@@ -35,11 +35,16 @@ export class ReviewPageComponent {
     }
 
     deleteReview(id: string) {
-      this.api.deleteReview(id).subscribe((res) => {
-        //refresh page
-        this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
-          this.router.navigate(['/reviews']);
+      const agree = confirm('Are you sure you want to delete this review?');
+      if(agree) {
+        this.api.deleteReview(id).subscribe((res) => {
+          //refresh page
+          this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
+            this.router.navigate(['/reviews']);
+          });
         });
-      });
+      } else {
+        return;
+      }
     }
 }
